+++ tests/testthat/test-utils.R
@@ -230,6 +230,35 @@ test_that("as_slide_computation raises errors as expected", {
   expect_error(as_time_slide_computation(5),
     class = "epiprocess__as_slide_computation__cant_convert_catchall"
   )
+
+  # If `.f` doesn't look like tidyeval and we fail to force it, then we hint to
+  # the user some potential problems:
+  toy_edf <- tibble(geo_value = 1, time_value = c(1, 2), value = 1:2) %>%
+    as_epi_df(as_of = 1)
+  toy_archive <- tibble(version = c(1, 2, 2), geo_value = 1, time_value = c(1, 1, 2), value = 1:3) %>%
+    as_epi_archive()
+  expect_error(
+    toy_edf %>%
+      group_by(geo_value) %>%
+      epi_slide(.window_size = 6, tibble(slide_value = mean(.x$value))),
+    class = "epiprocess__as_slide_computation__error_forcing_.f"
+  )
+  expect_snapshot(
+    error = TRUE,
+    toy_edf %>%
+      group_by(geo_value) %>%
+      epi_slide(.window_size = 6, tibble(slide_value = mean(.x$value)))
+  )
+  expect_error(
+    toy_archive %>%
+      epix_slide(tibble(slide_value = mean(.x$value))),
+    class = "epiprocess__as_slide_computation__error_forcing_.f"
+  )
+  expect_snapshot(
+    error = TRUE,
+    toy_archive %>%
+      epix_slide(tibble(slide_value = mean(.x$value)))
+  )
 })
 
 test_that("as_slide_computation works", {